Move snippets to the local folder
