+++ docs/csharp/fundamentals/types/records.md
@@ -1,7 +1,7 @@
 ---
 title: "Records in C#"
 description: Learn about C# record types and how to create them. A record is a class that provides value semantics.
-ms.date: 09/30/2021
+ms.date: 05/24/2023
 helpviewer_keywords: 
   - "records [C#]"
   - "C# language, records"
